Add possibility to sort call of addons
- Add notion of tag on addon. It allows to identify a group of addon if
they share the same tag
- Add the notion of EngineAddonSortingRule that addon can declare in
order to indicates if for a given engine event the current addon mus be
called before or after the addon with the given tag
- use AddonSortingRule to properly order EngineStatusView and
ModelAnimator addons with regards to ModelDebug Addon. This ensures that
these view are refreshed before the ModelDebugger pauses the execution.


contributes to
https://github.com/eclipse/gemoc-studio-modeldebugging/issues/111

Signed-off-by: Didier Vojtisek <didier.vojtisek@inria.fr>